# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Voucher API Documentation",
        "description": "API documentation for Voucher Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/tickets": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Create a new ticket",
                "operationId": "93882881cca9046d8c6ddbbab7309b4b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "subject",
                                    "message"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "MC34234"
                                    },
                                    "subject": {
                                        "type": "string",
                                        "example": "Account Issue"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "I cannot access my account"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket created successfully"
                                        },
                                        "ticket": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "ticket_id": {
                                                    "type": "string",
                                                    "example": "geg432"
                                                },
                                                "user_id": {
                                                    "type": "string",
                                                    "example": "MC34234"
                                                },
                                                "subject": {
                                                    "type": "string",
                                                    "example": "Account Issue"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "I cannot access my account"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "open"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/user/{user_id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get user's tickets",
                "operationId": "820120e6cfb0aeac2c4aa792174aa26d",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user's tickets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tickets": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "ticket_id": {
                                                        "type": "string",
                                                        "example": "geg432"
                                                    },
                                                    "user_id": {
                                                        "type": "string",
                                                        "example": "MC34234"
                                                    },
                                                    "subject": {
                                                        "type": "string",
                                                        "example": "Account Issue"
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "I cannot access my account"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "open"
                                                    },
                                                    "admin_reply": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "admin_replied_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}